fix: move canModifyProfile into profileRoutes and only admin and volunteer with own id can edit profile
